MANY THINGS
Reorginized stuff to be in a bin's folder first.
Added a shared folder for modules that all use the same thing (classic is a big one)
Added some basics for an interface to wit ai.
Added libcurl-x64 near main for luajit request. It's like socket except it works with SSL
the cert is also for that as well